requirements.txt
torchsde==0.2.6
einops==0.8.0
transformers==4.51.3
safetensors==0.4.3
accelerate==0.32.1
pyyaml==6.0.1
pillow==10.4.0
scipy==1.14.0
tqdm==4.66.4
psutil==6.0.0
pytorch_lightning==2.3.3
omegaconf==2.3.0
gradio==3.41.2
pygit2==1.15.1
opencv-contrib-python-headless==4.10.0.84
httpx==0.27.0
onnxruntime==1.18.1
timm==1.0.7
numpy==1.26.4
tokenizers==0.21.1
packaging==24.1
rembg==2.0.57
groundingdino-py==0.4.0
segment_anything==1.0
fastapi
pydantic==2.4.2
pydantic_core==2.10.1
python-multipart==0.0.6
uvicorn[standard]==0.23.2
chardet
colorlog
requests
numpy
sqlalchemy
packaging
rich
